Move HostOperation enum to objects
It semantically makes more sense for it to live next to the actual Host
object, which does carry the implementation of these operations.
