+++ tests/test_inventory.py
@@ -4,8 +4,8 @@ import pytest
 
 from exosphere.config import Configuration
 from exosphere.data import HostState
-from exosphere.inventory import FilterMode, HostOperation, Inventory, SortField
-from exosphere.objects import Host
+from exosphere.inventory import FilterMode, Inventory, SortField
+from exosphere.objects import Host, HostOperation
 from exosphere.security import SudoPolicy
 
 
@@ -35,42 +35,6 @@ class TestSortField:
     @pytest.mark.parametrize("field", list(SortField))
     def test_reverse_lookup_by_value(self, field):
         assert SortField(field.value) is field
-
-
-class TestHostOperation:
-    """
-    Tests for the HostOperation enum.
-
-    Since the enum carries extra semantics and has a constructor, we
-    ensure the basic enum properties hold and that the extra semantics
-    are valid, and correct.
-    """
-
-    @pytest.mark.parametrize("op", list(HostOperation))
-    def test_value_maps_to_callable_host_method(self, op):
-        """Test that each member's value is a callable on Host."""
-        assert callable(getattr(Host, op.value, None)), (
-            f"{op.name} -> '{op.value}' is not a callable Host method"
-        )
-
-    @pytest.mark.parametrize("op", list(HostOperation))
-    def test_has_label(self, op):
-        """Test that each member has a display label"""
-        assert isinstance(op.label, str) and op.label
-
-    @pytest.mark.parametrize("op", list(HostOperation))
-    def test_reverse_lookup_by_value(self, op):
-        """Test that each member can be looked up by its method-name value."""
-        assert HostOperation(op.value) is op
-
-    @pytest.mark.parametrize("op", list(HostOperation))
-    def test_modifies_state_is_bool(self, op):
-        """Each member exposes a boolean modifies_state flag."""
-        assert isinstance(op.modifies_state, bool)
-
-    def test_sync_does_not_modify_state(self):
-        """Sync should be stateless always"""
-        assert HostOperation.SYNC.modifies_state is False
 
 
 class TestInventory:
